Disallow !reset in any channel except #newcomers unless forced

+++ addons/mod.py
@@ -167,8 +167,12 @@ class Mod:
 
     @commands.has_permissions(kick_members=True)
     @commands.command(pass_context=True, name="reset")
-    async def reset(self, ctx, limit: int = 100):
+    async def reset(self, ctx, limit: int = 100, force: str = ""):
         """Wipes messages in #newcomers and pastes the welcome message again. Staff only."""
+        if(ctx.message.channel != self.bot.welcome_channel and force != "force"):
+            await self.bot.say("This command is limited to <#{}>".format(self.bot.welcome_channel.id))
+            return
+
         try:
             await self.bot.purge_from(ctx.message.channel, limit=limit)
